# api.yaml
openapi: 3.1.0
info:
  title: SegmentedEncoder rest api
  version: 0.0.3
servers:
  - description: Local
    url: http://localhost:8888/api
components:
  parameters:
    job_id:
      name: job_id
      required: true
      in: path
      schema:
        type: string
    task_id:
      name: task_id
      required: true
      in: path
      schema:
        type: string
  schemas:
    task_input:
      type: array
      items:
        type: object
        required:
          - input
        properties:
          input:
            type: integer
            minimum: 0
          start:
            type: number
            format: double
          end:
            type: number
            format: double
        example:
          - input: 0
            start: 0
            end: 2.5
    codec_params:
      type: object
      properties:
        codec:
          type: string
        params:
          type: array
          items:
            type: string
      example:
        codec: libaom-av1
        params: [ -crf, 30 ]
    task:
      type: object
      required: [ job_id, task_id, job_options, input, recipe ]
      properties:
        job_id:
          type: string
        task_id:
          type: string
        job_options:
          $ref: "#/components/schemas/job_options"
        input:
          $ref: "#/components/schemas/task_input"
        recipe:
          $ref: "#/components/schemas/recipe"
    transcode_task:
      type: object
      required: [ options ]
      properties:
        options:
          description: Additional options
          type: array
          items:
            type: string
    analysis_task:
      type: object
      properties:
        duration:
          type: number
          format: double
    merge_task:
      description: How to merge the inputs
      type: object
      required: [ concatenate ]
      properties:
        concatenate:
          type: array
          items:
            type: integer
    recipe:
      type: object
      properties:
        analysis:
          $ref: "#/components/schemas/analysis_task"
        transcode:
          $ref: "#/components/schemas/transcode_task"
        merge:
          $ref: "#/components/schemas/merge_task"
      minProperties: 1
      maxProperties: 1
      additionalProperties: false
    task_request:
      type: object
      required: [ inputs, recipe ]
      properties:
        inputs:
          $ref: "#/components/schemas/task_input"
        recipe:
          oneOf:
            - $ref: "#/components/schemas/transcode_task"
            - $ref: "#/components/schemas/merge_task"
    task_status:
      type: object
      properties:
        successfully_completed:
          type: boolean
    job_options:
      type: object
      required: [ video ]
      properties:
        video:
          $ref: "#/components/schemas/codec_params"
          required: [ codec ]
        audio:
          $ref: "#/components/schemas/codec_params"
      example:
        video:
          codec: "libsvtav1"
          params: [ "-crf", "30" ]
        audio:
          codec: "libopus"
  securitySchemes:
    auth_token:
      name: Authorization
      in: header
      type: apiKey
tags:
  - name: default
  - name: nice
    description: Nice optional features that arent needed for the software to work properly
  - name: worker
    description: Endpoints used by workers

paths:
  /version:
    description: get server/api version
    get:
      responses:
        200:
          description: Server version
          content:
            application/json:
              schema:
                type: string
              example: v1.2.3
  /login:
    description: Provide credentials to generate a session token
    get:
      parameters:
        - name: credentials
          in: header
          required: true
          schema:
            type: string
      responses:
        200:
          description: the authentication token for the created session
          content:
            application/json:
              schema:
                type: string
        403:
          description: Bad credentials
          content:
            text/plain:
              schema:
                type: string
              example: Bad credentials

  /job:
    get:
      description: List jobs
      security:
        - auth_token: [ ]
      responses:
        200:
          description: Array with job ids
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
    post:
      description: Create new job
      security:
        - auth_token: [ ]
      parameters:
        - name: video_codec
          required: true
          in: header
          example: libsvtav1
          schema:
            type: string
        - name: video_param
          in: header
          example: [ "-preset", "4", "-crf", "27" ]
          schema:
            type: array
            items:
              type: string
        - name: audio_codec
          in: header
          example: libopus
          schema:
            type: string
        - name: audio_param
          in: header
          example: [ "-b:a", "96k" ]
          schema:
            type: array
            items:
              type: string
        - name: segment_duration
          in: header
          description: Size of generated segmentes in seconds, 0 means single segment
          example: 25
          schema:
            type: number
            minimum: 0
      requestBody:
        required: true
        content:
          video/*:
            schema:
              type: string
              format: binary
      responses:
        201:
          description: Job created
          content:
            application/json:
              schema:
                type: string
  /job/{job_id}/info:
    get:
      description: Get job info, as well as each task progress
      security:
        - auth_token: [ ]
      parameters:
        - $ref: "#/components/parameters/job_id"
      responses:
        200:
          description: Job info/status
  /job/{job_id}/output:
    get:
      description: Get the job output (output of the last task)
      security:
        - auth_token: [ ]
      parameters:
        - $ref: "#/components/parameters/job_id"
      responses:
        200:
          description: The job output
        503:
          description: The job is not finished yet
  /job/{job_id}:
    delete:
      description: Delete this job and its tasks
      security:
        - auth_token: [ ]
      parameters:
        - $ref: "#/components/parameters/job_id"
      responses:
        204:
          description: Task instance canceled
  /job/{job_id}/task/{task_id}:
    delete:
      description: Abort task instance. Will rerun
      security:
        - auth_token: [ ]
      parameters:
        - $ref: "#/components/parameters/job_id"
        - $ref: "#/components/parameters/task_id"
      responses:
        204:
          description: Task instance canceled
        409:
          description: Task can not be canceled. Will happens after the task finishes
  /job/{job_id}/input:
    get:
      description: Get the job input
      security:
        - auth_token: [ ]
      parameters:
        - $ref: "#/components/parameters/job_id"
      responses:
        200:
          description: The job input
        404:
          description: The job not found
      tags:
        - nice
  #Workers endpoints
  /allocate_task:
    get:
      security:
        - auth_token: [ ]
      description: allocate a task to current worker
      responses:
        200:
          description: The allocated task info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/task"
        503:
          description: Timed out and no task was available, try again later
          content:
            text/plain:
              example: No task available, try again later
          headers:
            Retry-After:
              schema:
                type: integer
      tags:
        - worker
  /job/{job_id}/task:
    post:
      security:
        - auth_token: [ ]
      description: Append task to job
      parameters:
        - $ref: "#/components/parameters/job_id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/task_request"
      tags:
        - worker
      responses:
        201:
          description: Created task
  /job/{job_id}/task/{task_id}/input/{num}:
    get:
      security:
        - auth_token: [ ]
      parameters:
        - name: num
          in: path
          required: true
          schema:
            type: number
          example: 0
        - $ref: "#/components/parameters/job_id"
        - $ref: "#/components/parameters/task_id"
      tags:
        - worker
      responses:
        200:
          description: Task input file content
  /job/{job_id}/task/{task_id}/output:
    put:
      security:
        - auth_token: [ ]
      parameters:
        - $ref: "#/components/parameters/job_id"
        - $ref: "#/components/parameters/task_id"
      tags:
        - worker
      requestBody:
        content:
          video/x-matroska:
            schema:
              type: string
              format: binary
          audio/x-matroska:
            schema:
              type: string
              format: binary
      responses:
        201:
          description: Successfully uploaded task output
    get:
      security:
        - auth_token: [ ]
      parameters:
        - $ref: "#/components/parameters/job_id"
        - $ref: "#/components/parameters/task_id"
      tags:
        - worker
      responses:
        200:
          description: Task output
        404:
          description: Task output not available
  /job/{job_id}/task/{task_id}/status:
    post:
      security:
        - auth_token: [ ]
      description: Update task status
      parameters:
        - $ref: "#/components/parameters/job_id"
        - $ref: "#/components/parameters/task_id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/task_status"
      tags:
        - worker
      responses:
        200:
          description: Task status